// client/ionic/src/api.tsx
import {apiBaseUrl,elasticsearchHost,elasticsearchIndex} from './constants';

const API = {
  searchEndpoint() {
    return elasticsearchHost + '/' + elasticsearchIndex +'/';
  },
  loginUser() {
    return apiBaseUrl + 'user/login';
  },
  registerUser() {
    return apiBaseUrl + 'user/register';
  },
  entityIndexByType(entityType: string) {
    return apiBaseUrl + 'entity/index/' + entityType;
  },
  entityCreateByType(entityType: string) {
    return apiBaseUrl + 'entity/create/' + entityType;
  },
  entityUpdateByType(entityType: string, entityId: string) {
    return apiBaseUrl + 'entity/update/' + entityType + "/" + entityId;
  },
  entityGetByType(entityType: string, entityId: string) {
    return apiBaseUrl + 'entity/entity/' + entityType + '/' + entityId;
  },
  entityDetailsByType(entityType: string, entityId: string) {
    return apiBaseUrl + 'entity/details/' + entityType + '/' + entityId;
  },
  entitySchemaByType(entityType: string) {
    return apiBaseUrl + 'autogenerated/formSchema/' + entityType;
  }
};

export default API;