// client/ionic/src/services/web2pyrestful/api.ts
export interface InventoryItem {
    /**
     * 
     * @type {string}
     * @memberof InventoryItem
     */
    category?: string;
    /**
     * 
     * @type {string}
     * @memberof InventoryItem
     */
    idSeller?: string;
    /**
     * 
     * @type {boolean}
     * @memberof InventoryItem
     */
    isPublic?: boolean;
    /**
     * 
     * @type {string}
     * @memberof InventoryItem
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof InventoryItem
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof InventoryItem
     */
    image?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof InventoryItem
     */
    tags?: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof InventoryItem
     */
    availableQuantity?: number;
    /**
     * 
     * @type {number}
     * @memberof InventoryItem
     */
    id?: number;
}
/**
 * 
 * @export
 * @interface PurchaseOrder
 */
export interface PurchaseOrder {
    /**
     * 
     * @type {number}
     * @memberof PurchaseOrder
     */
    id?: number;
    /**
     * 
     * @type {Array<InventoryItem>}
     * @memberof PurchaseOrder
     */
    inventoryItems?: Array<InventoryItem>;
    /**
     * 
     * @type {string}
     * @memberof PurchaseOrder
     */
    paymentStatus?: string;
}